stream-processor/requirements.txt
confluent-kafka==2.4.0
pydantic==2.7.1
pydantic-settings==2.2.1
redis==5.0.4
psycopg2-binary==2.9.9
structlog==24.1.0
python-dotenv==1.0.1
